New geometries and macros added
Several geometries were implemented to test different situations:
Lead, Iron, and Uranium blocks of different widths.
A new results.C macro was created. This macro is used to study a lot
of observables: angles, spatial coordinates, time, charge,...and it is
used to permorf tomographic studies.

The TraRpcHit and TraRPC classes were corrected to follow the standard
Tragaldabas dimensions.

The TraBoxGenerator was modified. I changed the position of the generator
plane. Now the particles are generated towards the positive z-coordinate.